+++ app/store/chatList.go
@@ -8,14 +8,18 @@ import (
 	"time"
 
 	"github.com/nanu-c/axolotl/app/helpers"
+	"github.com/signal-golang/textsecure"
 	log "github.com/sirupsen/logrus"
 )
 
+// Session defines how a session looks like
 type Session struct {
 	ID           int64
+	UUID         string `db:"uuid"`
 	Name         string
 	Tel          string
 	IsGroup      bool
+	Type         int64 //describes the type of the session, wether it's a private conversation or groupv1 or groupv2
 	Last         string
 	Timestamp    uint64
 	When         string
@@ -28,7 +32,7 @@ type Session struct {
 	ExpireTimer  uint32 `db:"expireTimer"`
 }
 type MessageList struct {
-	ID       string
+	ID       int64
 	Session  *Session
 	Messages []*Message
 }
@@ -39,6 +43,11 @@ type Sessions struct {
 	Type       string
 }
 
+const (
+	invalidSession = -1
+	invalidQuote   = -1
+)
+
 //TODO that hasn't to  be in the db controller
 var AllSessions []*Session
 
@@ -46,57 +55,64 @@ var SessionsModel = &Sessions{
 	Sess: make([]*Session, 0),
 }
 
-func SaveSession(s *Session) error {
+// SaveSession saves a newly created session in the database
+func SaveSession(s *Session) (*Session, error) {
 	res, err := DS.Dbx.NamedExec(sessionsInsert, s)
 	if err != nil {
-		return err
+		return nil, err
 	}
 
 	id, err := res.LastInsertId()
 	if err != nil {
-		return err
+		return nil, err
 	}
 
 	s.ID = id
-	return err
+	return s, err
 }
+
+// UpdateSession updates a session in the database
 func UpdateSession(s *Session) error {
-	_, err := DS.Dbx.NamedExec("UPDATE sessions SET name = :name, timestamp = :timestamp, ctype = :ctype, last = :last, unread = :unread, notification = :notification, expireTimer = :expireTimer WHERE id = :id", s)
+	_, err := DS.Dbx.NamedExec("UPDATE sessions SET name = :name, timestamp = :timestamp, ctype = :ctype, last = :last, unread = :unread, notification = :notification, expireTimer = :expireTimer, uuid = :uuid WHERE id = :id", s)
 	if err != nil {
 		return err
 	}
 	return err
 }
-func DeleteSession(tel string) error {
-	s := SessionsModel.Get(tel)
-	_, err := DS.Dbx.Exec("DELETE FROM messages WHERE sid = ?", s.ID)
+
+// DeleteSession deletes a session in the database
+func DeleteSession(ID int64) error {
+	_, err := DS.Dbx.Exec("DELETE FROM messages WHERE sid = ?", ID)
 	if err != nil {
 		return err
 	}
-	_, err = DS.Dbx.Exec("DELETE FROM sessions WHERE id = ?", s.ID)
+	_, err = DS.Dbx.Exec("DELETE FROM sessions WHERE id = ?", ID)
 	if err != nil {
 		return err
 	}
-	index := SessionsModel.GetIndex(s.Tel)
-	SessionsModel.Sess = append(SessionsModel.Sess[:index], SessionsModel.Sess[index+1:]...)
+	SessionsModel.Sess = append(SessionsModel.Sess[:ID], SessionsModel.Sess[ID+1:]...)
 	SessionsModel.Len--
-	//qml.Changed(SessionsModel, &SessionsModel.Len)
 	return nil
 }
-func (s *Sessions) GetSession(i int) *Session {
+
+// GetSession at a certain index
+func (s *Sessions) GetSession(i int64) *Session {
 	return s.Sess[i]
 }
-func (s *Sessions) GetMessageList(id string) (error, *MessageList) {
-	index := SessionsModel.GetIndex(id)
-	if index == -1 {
-		index = int(SessionsModel.Get(id).ID)
-	}
-	if index != -1 {
-		messageList := &MessageList{
-			ID:      id,
-			Session: s.GetSession(index),
+
+// GetMessageList returns the message list for the session id
+func (s *Sessions) GetMessageList(ID int64) (error, *MessageList) {
+	if ID != invalidSession {
+		sess, err := s.Get(ID)
+		if err != nil {
+			log.Errorln("[axolotl] get messagelist", err)
+			return err, nil
 		}
-		err := DS.Dbx.Select(&messageList.Messages, messagesSelectWhere, messageList.Session.ID)
+		messageList := &MessageList{
+			ID:      ID,
+			Session: sess,
+		}
+		err = DS.Dbx.Select(&messageList.Messages, messagesSelectWhere, ID)
 		if err != nil {
 			log.Errorln("[axolotl] get messagelist", err)
 			return err, nil
@@ -104,7 +120,7 @@ func (s *Sessions) GetMessageList(id string) (error, *MessageList) {
 		// attach the quoted messages
 		for i, m := range messageList.Messages {
 			if m.Flags == helpers.MsgFlagQuote {
-				if m.QuoteID != -1 {
+				if m.QuoteID != invalidQuote {
 					err, qm := GetMessageById(m.QuoteID)
 					if err != nil {
 						log.Debugln("[axolotl] messagelist quoted message: ", err)
@@ -120,19 +136,23 @@ func (s *Sessions) GetMessageList(id string) (error, *MessageList) {
 			return err, nil
 		}
 		return nil, messageList
-	} else {
-		return errors.New("wrong index"), nil
 	}
-
+	return errors.New("wrong index"), nil
 }
-func (s *Sessions) GetMoreMessageList(id string, lastId string) (error, *MessageList) {
-	index := SessionsModel.GetIndex(id)
-	if index != -1 {
-		messageList := &MessageList{
-			ID:      id,
-			Session: s.GetSession(index),
+
+// GetMoreMessageList loads more messages from before the lastID
+func (s *Sessions) GetMoreMessageList(ID int64, lastID string) (error, *MessageList) {
+	if ID != -1 {
+		sess, err := s.Get(ID)
+		if err != nil {
+			log.Errorln("[axolotl] get messagelist", err)
+			return err, nil
 		}
-		err := DS.Dbx.Select(&messageList.Messages, messagesSelectWhereMore, messageList.Session.ID, lastId)
+		messageList := &MessageList{
+			ID:      ID,
+			Session: sess,
+		}
+		err = DS.Dbx.Select(&messageList.Messages, messagesSelectWhereMore, messageList.Session.ID, lastID)
 		if err != nil {
 			fmt.Println(err)
 			return err, nil
@@ -152,12 +172,12 @@ func (s *Sessions) GetMoreMessageList(id string, lastId string) (error, *Message
 			}
 		}
 		return nil, messageList
-	} else {
-		return errors.New("wrong index"), nil
 	}
-
+	return errors.New("wrong index"), nil
 }
-func (s *Session) Add(text string, source string, file []Attachment, mimetype string, outgoing bool, sessionID string) *Message {
+
+// Add message to a session
+func (s *Session) Add(text string, source string, file []Attachment, mimetype string, outgoing bool, sessionID int64) *Message {
 	var files []Attachment
 
 	ctype := helpers.ContentTypeMessage
@@ -168,7 +188,7 @@ func (s *Session) Add(text string, source string, file []Attachment, mimetype st
 			files = append(files, Attachment{File: fi.File, CType: ctype, FileName: fi.FileName})
 		}
 	}
-	fJson, err := json.Marshal(files)
+	fJSON, err := json.Marshal(files)
 	if err != nil {
 		log.Errorln(err)
 	}
@@ -178,7 +198,7 @@ func (s *Session) Add(text string, source string, file []Attachment, mimetype st
 		Outgoing:   outgoing,
 		Source:     source,
 		CType:      ctype,
-		Attachment: string(fJson),
+		Attachment: string(fJSON),
 		HTime:      "Now",
 		SentAt:     uint64(time.Now().UnixNano() / 1000000),
 	}
@@ -186,11 +206,8 @@ func (s *Session) Add(text string, source string, file []Attachment, mimetype st
 	s.Last = text
 	s.Len++
 	s.CType = ctype
-	//qml.Changed(s, &s.Last)
-	//qml.Changed(s, &s.Len)
-	//qml.Changed(s, &s.CType)
-	//FIXME
-	if !outgoing && sessionID != s.Tel {
+	//FIXME not shure if it breaks unread message counter
+	if !outgoing {
 		s.Unread++
 		//qml.Changed(s, &s.Unread)
 	}
@@ -199,11 +216,14 @@ func (s *Session) Add(text string, source string, file []Attachment, mimetype st
 	s.moveToTop()
 	return message
 }
+
+// MarkRead marks a session as read
 func (s *Session) MarkRead() {
 	s.Unread = 0
-	//qml.Changed(s, &s.Unread)
 	UpdateSession(s)
 }
+
+// ToggleSessionNotifcation turns on/off notification for a session
 func (s *Session) ToggleSessionNotifcation() {
 	s.Notification = !s.Notification
 	txt := ""
@@ -218,9 +238,8 @@ func (s *Session) ToggleSessionNotifcation() {
 	UpdateSession(s)
 }
 
-// updateTimestamps keeps the timestamps of the last message of each session
+// UpdateTimestamps keeps the timestamps of the last message of each session
 // updated in human readable form.
-// FIXME: make this lazier, to only update timestamps the user sees at the moment
 func UpdateTimestamps() {
 	for {
 		time.Sleep(1 * time.Minute)
@@ -230,26 +249,78 @@ func UpdateTimestamps() {
 			}
 			for _, m := range s.Messages {
 				m.HTime = helpers.HumanizeTimestamp(m.SentAt)
-				//qml.Changed(m, &m.HTime)
 			}
 			s.When = s.Messages[len(s.Messages)-1].HTime
-			//qml.Changed(s, &s.When)
 		}
 	}
 }
-func (s *Sessions) Get(tel string) *Session {
+
+// Get returns the session by it's id
+func (s *Sessions) Get(id int64) (*Session, error) {
+	for _, ses := range s.Sess {
+		if ses.ID == id {
+			return ses, nil
+		}
+	}
+	return nil, fmt.Errorf("Session with id %d not found", id)
+}
+
+// GetByE164 returns the session by the telephone number and creates it if it doesn't exists
+func (s *Sessions) GetByE164(tel string) *Session {
 	for _, ses := range s.Sess {
 		if ses.Tel == tel {
 			return ses
 		}
 	}
-	ses := &Session{Tel: tel, Name: TelToName(tel), Active: true, IsGroup: tel[0] != '+', Notification: true}
+	newSession := s.CreateSessionForE164(tel, "0")
+	return newSession
+}
+
+// CreateSessionForE164 creates a new Session for the phone number
+func (s *Sessions) CreateSessionForE164(tel string, UUID string) *Session {
+	ses := &Session{Tel: tel,
+		Name:         TelToName(tel),
+		Active:       true,
+		IsGroup:      false,
+		Notification: true,
+		UUID:         UUID,
+	}
 	s.Sess = append(s.Sess, ses)
 	s.Len++
-	//qml.Changed(s, &s.Len)
 	SaveSession(ses)
 	return ses
 }
+
+// CreateSessionForGroup creates a session for a group
+func (s *Sessions) CreateSessionForGroup(group *textsecure.Group) *Session {
+	ses := &Session{Tel: group.Hexid, // for legacy reasons add group id also as Tel number
+		Name:         group.Name,
+		Active:       true,
+		IsGroup:      true,
+		Notification: true,
+		UUID:         group.Hexid,
+	}
+	s.Sess = append(s.Sess, ses)
+	s.Len++
+	ses, err := SaveSession(ses)
+	if err != nil {
+		log.Errorln("CreateSessionForGroup failed:", err)
+		return nil
+	}
+	return ses
+}
+
+// GetByUUID returns the session by the ChatUUID
+func (s *Sessions) GetByUUID(UUID string) (*Session, error) {
+	for _, ses := range s.Sess {
+		if ses.UUID == UUID {
+			return ses, nil
+		}
+	}
+	return nil, fmt.Errorf("Session with uuid %s not found", UUID)
+}
+
+// UpdateSessionNames updates the non groups with the name from the phone book
 func (s *Sessions) UpdateSessionNames() {
 	for _, ses := range s.Sess {
 		if ses.IsGroup == false {
@@ -257,34 +328,33 @@ func (s *Sessions) UpdateSessionNames() {
 			UpdateSession(ses)
 		}
 	}
-	//qml.Changed(&SessionsModel, &SessionsModel.Len)
 }
-func (s *Sessions) GetIndex(tel string) int {
+
+// GetIndex returns the current index of session ID
+func (s *Sessions) GetIndex(ID int64) int {
 	for i, ses := range s.Sess {
-		if ses.Tel == tel {
+		if ses.ID == ID {
 			return i
 		}
 	}
 	return -1
 }
 
-var topSession string
+var topSession int64
 
 func (s *Session) moveToTop() {
-	if topSession == s.Tel {
+	if topSession == s.ID {
 		return
 	}
 
-	index := SessionsModel.GetIndex(s.Tel)
+	index := SessionsModel.GetIndex(s.ID)
 	SessionsModel.Sess = append([]*Session{s}, append(SessionsModel.Sess[:index], SessionsModel.Sess[index+1:]...)...)
 
 	// force a length change update
 	SessionsModel.Len--
-	//qml.Changed(SessionsModel, &SessionsModel.Len)
 	SessionsModel.Len++
-	//qml.Changed(SessionsModel, &SessionsModel.Len)
 
-	topSession = s.Tel
+	topSession = s.ID
 }
 func LoadChats() error {
 	log.Printf("[axolotl] Loading Chats")
@@ -314,6 +384,5 @@ func LoadChats() error {
 			m.HTime = helpers.HumanizeTimestamp(m.SentAt)
 		}
 	}
-	//qml.Changed(SessionsModel, &SessionsModel.Len)
 	return nil
 }